+++ src/utils/getUpdateType.ts
@@ -3,7 +3,7 @@ import * as semver from "semver";
 export function getUpdateType(
   currentValue: string,
   newValue: string
-): "major" | "minor" | "patch" | "latest" | "out of range" {
+): "major" | "minor" | "patch" | "latest" {
 
   const isRange = currentValue.startsWith("^") || currentValue.startsWith("~");
   const isLatest = currentValue === "latest";
@@ -19,22 +19,6 @@ export function getUpdateType(
 
   if (!semver2) {
     throw new Error("Invalid semver string");
-  }
-
-  if (isRange) {
-    if (!semver.satisfies(semver2.version, currentValue)) {
-      return "out of range";
-    } else {
-      if (semver1!.major !== semver2.major) {
-        return "major";
-      } else if (semver1!.minor !== semver2.minor) {
-        return "minor";
-      } else if (semver1!.patch !== semver2.patch) {
-        return "patch";
-      } else {
-        return "latest";
-      }
-    } 
   }
 
   if (semver1!.major !== semver2.major) {