+++ CHANGELOG.md
@@ -4,6 +4,32 @@ All notable changes to the "update-now" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+
+## [0.0.5] - 2023-11-11
+
+### Added
+
+ - New tooltip when hovering the Code Lens
+ - Added author and description information to the tooltip
+
+### Changed
+
+- Updated copy in the readme
+- Cleaned code
+- Updated emojis to ✅ ⚠️ 🛑 for patch, minor, marjor update
+
+### Removed
+
+- Out of date dependencies count
+
+
+
+## [0.0.4] - 2023-11-06
+
+### Added
+
+- Saved file after the dependency is updated.
+  
 ## [0.0.3] - 2023-11-05
 
 ### Added
@@ -35,6 +61,3 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.0.1] - 2023-05-24
 
 - Initial release
-
-
-## [Unreleased]